+++ NEWS.md
@@ -1,3 +1,10 @@
-# PLreg 0.2.0.9000
+# PLreg 0.3.0
 
-* Added a `NEWS.md` file to track changes to the package.
+## Major changes
+
+* Correction of the hessian matrix in the PLreg function.
+* Added sandwich function.
+
+## Minor changes
+
+* Change in the first sentence of the description field of the DESCRIPTION file.